requirements.txt
fastapi
uvicorn
gunicorn
requests
python-dotenv
pydantic
pinecone-client
sentence-transformers